fix clippy warnings

+++ sudachi/src/dic/lexicon/trie.rs
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Works Applications Co., Ltd.
+ * Copyright (c) 2021-2024 Works Applications Co., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -100,7 +100,7 @@ impl<'a> Trie<'a> {
     }
 
     pub fn total_size(&self) -> usize {
-        4 * self.array.len() as usize
+        4 * self.array.len()
     }
 
     #[inline]
