+++ sudachi/src/hash.rs
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2021 Works Applications Co., Ltd.
+ *  Copyright (c) 2021-2024 Works Applications Co., Ltd.
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -91,7 +91,7 @@ impl Hasher for RoMuHash {
 
     #[inline(always)]
     fn write_u64(&mut self, i: u64) {
-        self.consume(i as u64);
+        self.consume(i);
     }
 
     #[inline(always)]
